requirements.txt
newsapi-python      0.2.6
newspaper3k         0.2.8
pyTelegramBotAPI    3.7.4
python-telegram-bot 13.6
telebot             0.0.4


